add `enabled` for sprite and label

+++ engine/gamesys/src/gamesys/components/comp_gui.cpp
@@ -2412,8 +2412,7 @@ namespace dmGameSystem
         }
         index -= num_world_properties;
 
-
-        uint64_t num_bool_properties = 1;
+        uint32_t num_bool_properties = 1;
         if (index < num_bool_properties)
         {
             if (index == 0)
